Fix Unity compilation errors for namespace conflicts
- Fix Physics namespace conflicts by using UnityEngine.Physics explicitly
- Fix memory calculation cast for long type conversion
- Resolve compilation errors preventing Unity script usage

Errors fixed:
- CS0234: Physics.Raycast namespace conflict
- CS0234: Physics.OverlapSphere namespace conflict
- CS0266: Implicit float to long conversion

Scripts now compile successfully in Unity 2022.3.50.

+++ Assets/Scripts/Utils/PerformanceMonitor.cs
@@ -122,7 +122,7 @@ namespace DiceGame.Utils
             currentFPS = 1.0f / Time.unscaledDeltaTime;
             
             // Get memory usage
-            currentMemoryMB = System.GC.GetTotalMemory(false) / (1024 * 1024);
+            currentMemoryMB = (long)(System.GC.GetTotalMemory(false) / (1024 * 1024));
             
             // Update history
             UpdateFPSHistory(currentFPS);
